+++ changelog.md
@@ -8,6 +8,9 @@ This version is not released yet and is under active development.
 
 - New `--show-params` option to debug parameters defaults, values, environment variables and provenance.
 - Rename `ignored_options` to `ignored_params` on `ConfigOption`.
+- Highlight command's metavars, default values and deprecated flag in help.
+- Finer highlighting of options, subcommands and their aliases in help screens.
+- Fix highlight of dynamic metavars and secondary option in help screen.
 - New custom `ExtraContext` which allows populating `meta` at instanciation.
 - Use the `Formats` enum to encode for default configuration file extensions.
 - Re-introduce `*.yml` as a possible extension for YAML files.